feat: finish module task
